GH-1071: JUnit 5 Support Improvements
Resolves https://github.com/spring-projects/spring-amqp/issues/1071

- Remove JUnit4 dependency from `RabbitAvailableCondition` (minor breaking API change)
- Add `purgeAfterEach` to `@RabbitAvailable`
- tabs not spaces in `RabbitAvailableCondition` (review with `?w=1`)
- `@LogLevels` now requires `level`
